small refactoring

+++ src/main/java/com/wine/to/up/user/service/controller/NotificationTokensController.java
@@ -1,6 +1,7 @@
 package com.wine.to.up.user.service.controller;
 
 import com.wine.to.up.commonlib.security.AuthenticationProvider;
+import com.wine.to.up.user.service.domain.dto.NotificationTokenDto;
 import com.wine.to.up.user.service.domain.entity.NotificationTokenType;
 import com.wine.to.up.user.service.service.NotificationTokensService;
 import io.swagger.annotations.Api;
@@ -18,13 +19,16 @@ import org.springframework.web.bind.annotation.*;
 @Slf4j
 @Api(value = "NotificationTokensController", description = "Operations with notification tokens")
 public class NotificationTokensController {
-    public final NotificationTokensService notificationTokensService;
+    private final NotificationTokensService notificationTokensService;
 
     @ApiOperation(value = "Add token to list",
             notes = "Description: Adds application token to notification tokens list of currently logged in user")
     @PostMapping(path = "/")
     @PreAuthorize("isAuthenticated()")
-    public ResponseEntity<Void> addUserNotificationToken(@RequestParam String token, @RequestParam NotificationTokenType tokenType) {
+    public ResponseEntity<Void> addUserNotificationToken(
+        @RequestParam String token,
+        @RequestParam NotificationTokenType tokenType
+    ) {
         notificationTokensService.addNotificationToken(AuthenticationProvider.getUser().getId(), token, tokenType);
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -32,9 +36,12 @@ public class NotificationTokensController {
     @PostMapping(path = "/{userId}")
     @ApiOperation(value = "Add token to list",
             notes = "Description: Adds application token to notification tokens list")
-
-    public ResponseEntity<Void> addUserNotificationTokenById(@PathVariable Long userId, @RequestParam String token,
-                                                         @RequestParam NotificationTokenType tokenType) {
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Void> addUserNotificationTokenById(
+        @PathVariable Long userId,
+        @RequestParam String token,
+        @RequestParam NotificationTokenType tokenType
+    ) {
         notificationTokensService.addNotificationToken(userId, token, tokenType);
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -42,7 +49,12 @@ public class NotificationTokensController {
     @ApiOperation(value = "Remove app token from list",
             notes = "Description: Removes application token from notification tokens list of currently logged in user")
     @DeleteMapping(path = "/")
+    @PreAuthorize("isAuthenticated()")
     public ResponseEntity<Void> removeNotificationToken(@RequestParam String token) {
+        final NotificationTokenDto byId = notificationTokensService.getById(token);
+        if (!byId.getUserId().equals(AuthenticationProvider.getUser().getId())) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         notificationTokensService.removeNotificationToken(token);
         return new ResponseEntity<>(HttpStatus.OK);
     }
